+++ app/_components/header/_component/header-not-login.tsx
@@ -8,7 +8,7 @@ interface HeaderLoginProps {
 
 export function HeaderNotLogin({ onClick }: HeaderLoginProps) {
   return (
-    <header className="fixed left-0 right-0 top-0 hidden md:flex">
+    <>
       <Image src="/images/logo.png" alt="Logo" width={133} height={43} />
       <nav>
         <ul>
@@ -21,10 +21,15 @@ export function HeaderNotLogin({ onClick }: HeaderLoginProps) {
             <Link href="/signup">회원가입</Link>
           </li>
           <li>
-            <Image src="/icons/noti-none.png" alt="notification" width={24} height={24} />
+            <Image
+              src="/icons/noti-none.png"
+              alt="notification"
+              width={24}
+              height={24}
+            />
           </li>
         </ul>
       </nav>
-    </header>
+    </>
   );
 }
